refactor: domain 객체 폴더 변경
