+++ src/main/java/step3/domain/RacingCar.java
@@ -1,4 +1,4 @@
-package step3.dto;
+package step3.domain;
 
 public class RacingCar {
