+++ src/main/java/step3/view/ResultView.java
@@ -1,6 +1,6 @@
 package step3.view;
 
-import step3.dto.RacingCar;
+import step3.domain.RacingCar;
 
 public class ResultView {
     public static void printResult(RacingCar[] racingCars) {
